+++ CHANGELOG.md
@@ -21,6 +21,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Support for tuple conversions with proper error handling and type safety
 - `TupleRow<Idx>` and `TupleRowMut<Idx>` traits for indexing rows in tuples of tuples
 - `FirstTupleRow` and `LastTupleRow` convenience traits for accessing the first and last rows in tuples of tuples
+- `FirstTupleIndex` and `LastTupleIndex` convenience traits for accessing the first and last elements in tuples
 
 ## [0.1.0] - 2025-12-05